requirement.txt
Flask==3.0.2
Flask-Cors==4.0.0
torch==2.2.1
diffusers==0.27.2
transformers==4.39.3
numpy==1.26.4
Pillow==10.2.0
google-generativeai==0.3.2
werkzeug==3.0.1
python-dotenv==1.0.1
accelerate==0.29.2